Fixes OPENSTACK-1106 Not a direct fix for this issue. The root cause is that the upstream delete the db entry before the agent actually deletes the listener and cleans up the vif port on vrs and we are seeing a racecondition when we do listener delete and loadbalancer delete back to back. We are seeing an issue in update_port where we try to process_port_sg_group on the port and by then lb_delete triggered delete_port has deleted the port and process_port_sg raises an exception that vport does not exist and we end up not cleaning up the port on vrs. Call process_port_sg in update_port only if the old and new requested port sg doesn't match. Since in the lbaas case, the sg doesn't change during listener_delete, process_port_sg is skipped and we won't hit the vport not exist case. But the actual fix needs to go upstream. But this fix is good to have even otherwise this issue.
Closes-Bug: OPENSTACK-1106

+++ nuage_neutron/plugins/nuage/plugin.py
@@ -849,6 +849,11 @@ class NuagePlugin(base_plugin.BaseNuagePlugin,
         session = context.session
         original_port = self.get_port(context, id)
         old_port = copy.deepcopy(original_port)
+        new_sg = (set(port['port'].get(ext_sg.SECURITYGROUPS)) if
+                  port['port'].get(ext_sg.SECURITYGROUPS) else set())
+        orig_sg = (set(original_port.get(ext_sg.SECURITYGROUPS)) if
+                   original_port.get(ext_sg.SECURITYGROUPS) else set())
+        sgids_diff = list(new_sg ^ orig_sg)
         old_dhcp_options = copy.deepcopy(old_port['extra_dhcp_opts'])
         for old_dhcp_option in old_dhcp_options:
             self._translate_dhcp_option(old_dhcp_option)
@@ -940,7 +945,8 @@ class NuagePlugin(base_plugin.BaseNuagePlugin,
 
         if (subnet_mapping
                 and subnet_mapping['nuage_managed_subnet'] is False):
-            if (delete_security_groups or has_security_groups):
+            if (delete_security_groups or (has_security_groups and
+                                           sgids_diff)):
                 # delete the port binding and process new sg binding
                 self._delete_port_security_group_bindings(context, id)
                 sgids = self._get_security_groups_on_port(context, port)
